openimageio: Fix the build with modern GCC
Newer GCC is stricter (for good reasons) and as such it reports more
warnings.

The OpenImageIO release we are bundling predates those new warnings, and
as such it is not ready for it.

This commit disables -Werror, so that the new warnings aren't fatal.

Updating to a newer OpenImageIO with the next release of Blender would
most likely remove the need for this.